AspectJ: cleanup and documentation
