+++ org.jacoco.core/src/org/jacoco/core/internal/analysis/filter/AspectJAroundClosureClassFilter.java
@@ -12,34 +12,32 @@
  *******************************************************************************/
 package org.jacoco.core.internal.analysis.filter;
 
-import org.objectweb.asm.Attribute;
 import org.objectweb.asm.tree.MethodNode;
 
+import java.util.regex.Pattern;
+
 /**
- * Filters methods with the {@code org.aspectj.weaver.AjSynthetic} attribute.
+ * Filters *$AjcClosure42 classes generated by the AspectJ compiler.
  */
-public class AspectJSyntheticFilter implements IFilter {
+public class AspectJAroundClosureClassFilter implements IFilter {
+
+	private static final Pattern AJC_CLOSURE_PATTERN = Pattern
+			.compile(".*\\$AjcClosure\\d+");
+
 	@Override
 	public void filter(MethodNode methodNode, IFilterContext context,
 			IFilterOutput output) {
 
-		if (isAjSynthetic(methodNode)) {
+		if (isAroundClosureClass(context)) {
 			output.ignore(methodNode.instructions.getFirst(),
 					methodNode.instructions.getLast());
 		}
 	}
 
-	public static boolean isAjSynthetic(MethodNode methodNode) {
-		if (methodNode.attrs == null) {
-			return false;
-		}
-
-		for (Attribute attr : methodNode.attrs) {
-			if (attr.type.equals("org.aspectj.weaver.AjSynthetic")) {
-				return true;
-			}
-		}
-
-		return false;
+	private boolean isAroundClosureClass(IFilterContext context) {
+		return context.getSuperClassName()
+				.equals("org/aspectj/runtime/internal/AroundClosure")
+				&& AJC_CLOSURE_PATTERN.matcher(context.getClassName())
+						.matches();
 	}
 }
