+++ org.jacoco.core/src/org/jacoco/core/internal/analysis/filter/AspectJInitFilter.java
@@ -15,6 +15,45 @@ package org.jacoco.core.internal.analysis.filter;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.tree.*;
 
+/**
+ * Filters calls to ajc$preClinit and ajc$postClinit inside the static
+ * initializer generated by the AspectJ compiler.
+ * <p>
+ * Example (decompiled) java code:
+ *
+ * <pre>{@code
+ * static {
+ * 	ajc$preClinit();
+ *
+ * 	try {
+ * 		ajc$postClinit();
+ * 	} catch (Throwable var1) {
+ * 		ajc$initFailureCause = var1;
+ * 	}
+ * }
+ * }</pre>
+ *
+ * Example javap output:
+ *
+ * <pre>
+ *   static {};
+ *     descriptor: ()V
+ *     flags: (0x0008) ACC_STATIC
+ *     Code:
+ *       stack=1, locals=1, args_size=0
+ *          0: invokestatic  #135                // Method ajc$preClinit:()V
+ *          3: nop
+ *          4: invokestatic  #73                 // Method ajc$postClinit:()V
+ *          7: goto          15
+ *         10: astore_0
+ *         11: aload_0
+ *         12: putstatic     #66                 // Field ajc$initFailureCause:Ljava/lang/Throwable;
+ *         15: return
+ *       Exception table:
+ *          from    to  target type
+ *              4    10    10   Class java/lang/Throwable
+ * </pre>
+ */
 public class AspectJInitFilter implements IFilter {
 	@Override
 	public void filter(MethodNode methodNode, IFilterContext context,
